fix(core-snapshots): don't append duplicate rounds rows to a snapshot (#2662)
When appending to an existent snapshot, only append rows with higher id
than already in the (old) snapshot. Otherwise the snapshot would contain
rows with duplciate rounds.id values and would not be importable.

This patch assumes no gaps in rounds.id and monotonically increasing
rounds.id when new rounds are being added.